use a seperate file to preprocess and devided files into different categories
